Rework Search
MERGE WITH FRONTEND-INIT
rebuild search for implement with branch frontend-init
