+++ Cinema_Project/Cinema_Project/Controllers/HomeController.cs
@@ -12,23 +12,15 @@ namespace Cinema_Project.Controllers
     [Authorize]
     public class HomeController : Controller
     {
-        private readonly ILogger<HomeController> _logger;
-        private readonly AppDbContext _context;
 
 
-        public HomeController(ILogger<HomeController> logger, AppDbContext context)
-        {
-            _logger = logger;
-            _context = context;
-        }
+
+
 
 
         public IActionResult Index()
         {
-            var genres = _context.Genres.ToList();
-            var movies = _context.Movies.Include(m => m.MovieGenres).ToList(); // Включаем связанные жанры
-            var viewModel = new MovieGenreViewModel { Movies = movies, Genres = genres };
-            return View(viewModel);
+            return View();
         }
 
 
@@ -45,40 +37,7 @@ namespace Cinema_Project.Controllers
         }
 
 
-        [HttpGet]
-        public IActionResult GetMovieDetails(int id)
-        {
-            var movie = _context.Movies.FirstOrDefault(m => m.MovieId == id);
-            return PartialView("_MovieDetailsPartial", movie);
-        }
 
-        [HttpGet]
-        public IActionResult Search(string search)
-        {
-            List<Movie> movies;
-
-            if (!string.IsNullOrEmpty(search))
-            {
-                movies = _context.Movies.Where(m => m.Title.Contains(search)).ToList();
-            }
-            else
-            {
-                movies = new List<Movie>();
-            }
-
-            return Json(movies);
-        }
-
-        [HttpGet]
-        public IActionResult FilterMovies(List<int> genres)
-        {
-            var movies = _context.Movies
-                .Where(m => m.MovieGenres.Any(g => genres.Contains(g.GenreId)))
-                .Include(m => m.MovieGenres)
-                .ToList();
-
-            return Json(movies);
-        }
