+++ Cinema_Project/Cinema_Project/Views/Shared/_Layout.cshtml
@@ -25,6 +25,9 @@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Search" asp-action="MainLogic">MainLogic</a>
+                        </li>
                     </ul>
                     <partial name="_LoginPartial" />
                 </div>
